Unprofessionally brisk upload to contribute
This is very unready for the public, but thanks to very helpful redditor Delicious_Pair_4828 for teaching me a real life example of how insanely useful functions are. Have to upload so I have a good reference, for me, not for you. Well maybe if you can learn from it it's for you!